refactored toml_to_py methods to use visitation and moving
also:
- added basic benchmark script with test data
- made profiling a compile-time option
- cleaned up code organization to separate definitions and implementations
- added profile scope to cover toml::parse
- updated toml++ to 2.1.0
- added .gitattributes to tell github not to count .toml files or third_party towards repository stats
