+++ .gitattributes
@@ -1,0 +1,3 @@
+# tell github to ignore these files for the purposes of language stats
+third_party/* linguist-vendored
+benchmark/*.toml linguist-vendored
