+++ benchmark/data.toml
@@ -1,0 +1,94 @@
+addition-used = false
+afraid-boiling-draconian = 1916-12-15T11:26:42+11:26
+boiling-whip = 12:23:17.000000129
+charming = 16771
+close-unbecoming = 14:49:02
+contain-terrible-neck = 1928-11-07T06:30:56
+creature = 0.4857
+crowded-profuse-capable = 1932-05-28
+deserted-cross-creature = 1931-12-26T04:31:04
+deserted-heavy = [ 2014-02-08, 1984-03-16, 1910-08-05 ]
+fabulous = 0.3388
+ghost-birds-legal = 1989-09-02
+grandfather-weary = 10216
+haircut-dreary-broken = 'unadvised rejoice lewd crime grandfather rice imperfect'
+harm-snow = 0.6724
+leather-aromatic-rabbit = false
+license-harm = 1987-08-15T15:30:51+03:24
+naughty-self-close = 1916-09-21T01:53:27-02:59
+neck = 29653
+neck-crime-wistful = 1998-12-07
+rest-afraid = [
+    1912-06-23T07:00:07-07:46,
+    1910-11-17T13:10:34,
+    2014-08-03T04:37:12-06:36,
+    1966-11-22T18:08:23+06:10,
+    1985-03-01T08:11:38,
+    1976-09-04T20:46:46+00:29,
+    1964-05-18T20:13:27-01:25,
+    1961-10-14T11:03:51.000011555+06:23,
+    2005-07-13T02:57:15-01:46,
+    2007-02-01T17:09:54-08:45,
+    1952-03-06T09:15:04+01:44,
+    1962-12-25T23:00:11+04:28
+]
+rotten-best = 0.0325
+run-curious = true
+run-meaty = 04:40:20.000003312
+scintillating-cream = 0.2604
+sisters-playground = 'week spiritual pause insidious uptight'
+spiritual-borrow-messy = 16:49:32
+surprise-jail-babies = 2008-01-01
+title-unbiased = 1958-10-08
+unbecoming-scene-lewd = true
+worried-spicy = 23:03:54
+
+[contain]
+ghost = 0.2471
+tank-lackadaisical = 1945-02-18T22:22:27+01:41
+
+    [contain.righteous-imperfect-remarkable]
+    ticket = 0.2412
+    voyage = 'prefer silky night draconian three dreary decisive'
+    wilderness-draconian = 04:58:17
+
+        [contain.righteous-imperfect-remarkable.deeply-night-gold]
+        legal = 1981-07-04
+        righteous = [ false, true, false, false, true ]
+        wilderness = 1900-12-24
+
+[crime]
+boiling-excuse-incandescent = 31173
+flat-rejoice = 0.8682
+furtive = 6866
+heavy-memory = 5545
+lackadaisical-terrible-overjoyed = [ 'harm acceptable naughty pause', 'page anxious threatening lewd' ]
+reject-aromatic = [ false, true, true, true, false, true ]
+reject-naughty = 11229
+scintillating-bake-harm = 0.8258
+
+[license-bat]
+adorable-blind-string = 1961-08-14
+bat = 1939-10-28
+blind-wilderness = 2004-07-14
+
+[rapid]
+borrow = true
+jail-weary-furtive = 'room'
+protective-sulky = 4737
+room-decisive-unbiased = 1998-11-01T12:23:58
+soap = 1985-05-06
+sprout-rich = 0.6026
+
+    [rapid.clean-leather]
+    blood-snow-mark = true
+    equable = 0.2923
+    fabulous = 1979-07-24T00:53:39+06:22
+    glorious-threatening = true
+    lewd-dull = 02:40:53.000027288
+
+[run-overjoyed]
+creature-cheerful = 11071
+respect-history-anxious = 0.4637
+vessel = 'curious'
+
