+++ benchmark/requirements.txt
@@ -1,0 +1,2 @@
+toml
+tomlkit
